fixup! refactor: move logger to the root
