+++ python/beeai_framework/memory/file_cache.py
@@ -21,10 +21,10 @@ from typing import Any, Generic, Self, TypeVar
 
 import aiofiles
 
+from beeai_framework.logger import BeeLogger
 from beeai_framework.memory.base_cache import BaseCache
 from beeai_framework.memory.serializer import Serializer
 from beeai_framework.memory.sliding_cache import SlidingCache
-from beeai_framework.utils import BeeLogger
 
 logger = BeeLogger(__name__)
