+++ python/beeai_framework/tools/weather/openmeteo.py
@@ -25,9 +25,9 @@ from pydantic import BaseModel, Field
 
 from beeai_framework.context import RunContext
 from beeai_framework.emitter.emitter import Emitter
+from beeai_framework.logger import BeeLogger
 from beeai_framework.tools import ToolInputValidationError
 from beeai_framework.tools.tool import StringToolOutput, Tool, ToolRunOptions
-from beeai_framework.utils import BeeLogger
 
 logger = BeeLogger(__name__)
