+++ python/beeai_framework/utils/__init__.py
@@ -13,8 +13,7 @@
 # limitations under the License.
 
 
-from beeai_framework.logger import BeeLogger
 from beeai_framework.utils.config import CONFIG
 from beeai_framework.utils.events import MessageEvent
 
-__all__ = ["CONFIG", "BeeLogger", "MessageEvent"]
+__all__ = ["CONFIG", "MessageEvent"]
