+++ python/examples/logger/agent.py
@@ -7,8 +7,8 @@ from beeai_framework.agents.bee.agent import BeeAgent
 from beeai_framework.agents.types import BeeRunOutput
 from beeai_framework.backend.chat import ChatModel
 from beeai_framework.errors import FrameworkError
+from beeai_framework.logger import BeeLogger
 from beeai_framework.memory.unconstrained_memory import UnconstrainedMemory
-from beeai_framework.utils import BeeLogger
 
 
 async def main() -> None:
