+++ python/examples/tools/decorator.py
@@ -10,9 +10,9 @@ from beeai_framework import BeeAgent, tool
 from beeai_framework.agents.types import AgentExecutionConfig
 from beeai_framework.backend.chat import ChatModel
 from beeai_framework.errors import FrameworkError
+from beeai_framework.logger import BeeLogger
 from beeai_framework.memory.unconstrained_memory import UnconstrainedMemory
 from beeai_framework.tools.tool import StringToolOutput
-from beeai_framework.utils import BeeLogger
 
 logger = BeeLogger(__name__)
